Added new code for management of rational approximations. A database of r_app is used to find the best approximation during a run. The database can be generated with the program found in Update/remez/ (gmp required). A sample set of approximations for nf=1 runs is included for spectral interval [0.0004:64].
Corrected a bug in local_action declaration.


git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/HiRep@8 37a69121-b030-0410-a54a-fdca83ffee47

+++ Include/update.h
@@ -48,7 +48,7 @@ typedef enum {
  * H = | momenta |^2 + S_g + < phi1, phi2>
  */
 void local_hmc_action(local_action_type type,
-                      float *loc_action,
+                      double *loc_action,
                       suNg_algebra_vector *momenta,
                       suNf_spinor *phi1,
                       suNf_spinor *phi2);
